requirements.txt
alabaster==0.7.13
aniso8601==9.0.1
annotated-types==0.5.0
anyio==4.0.0rc1
attrs==23.1.0
Babel==2.12.1
bjoern==3.2.2
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
docutils==0.20.1
exceptiongroup==1.1.2
flake8==6.0.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.0
flask-restx==1.1.0
flask-swagger-ui==4.11.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
humanize==4.7.0
idna==3.4
imagesize==1.4.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mccabe==0.7.0
mistune==3.0.1
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pptree==3.1
prettytable==3.8.0
psutil==5.9.5
pycodestyle==2.10.0
pydantic==2.1.1
pydantic_core==2.4.0
pyflakes==3.0.1
Pygments==2.15.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==3.5.3
referencing==0.30.0
rejson==0.5.6
requests==2.31.0
rpds-py==0.9.2
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.1.0
sphinx-click==4.4.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.27.0
tomli==2.0.1
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
